+++ internal/model/user.go
@@ -3,12 +3,12 @@ package model
 import "time"
 
 type User struct {
-	ID            int32     `db:"id" json:"id"`
+	ID            int64     `db:"id" json:"id"`
 	SiteBid       string    `db:"site_bid" json:"site_bid"`
-	ExtMemberID   int       `db:"ext_member_id" json:"ext_member_id"`
+	ExtMemberID   int64     `db:"ext_member_id" json:"ext_member_id"`
 	ExtUsername   string    `db:"ext_username" json:"ext_username"`
 	LastLoginTime time.Time `db:"last_login_time" json:"last_login_time"`
-	AvatarID      int       `db:"avatar_id" json:"avatar_id"`
+	AvatarID      int32     `db:"avatar_id" json:"avatar_id"`
 	IsSuspended   bool      `db:"is_suspended" json:"is_suspended"`
 	Code          string    `db:"code" json:"code"`
 	CreateTime    time.Time `db:"create_time" json:"create_time"`
